chore: finish removing openedx references for enterprise repos
As part of ENT-11240 (epic: ENT-11239) for cutting over enterprise IDAs
from the `openedx` github org to the `edx` github org, this change
codifies the cutover by updating devstack URLs to clone the correct URLs
and stop creating fork remotes.  This keeps things simple because we are
not actually making an effort to keep the forks in sync.